backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
supabase>=1.0.4
ultralytics>=8.0.0
onnxruntime>=1.24.0
pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
requests>=2.31.0
redis>=5.0.0
celery>=5.3.0
google-cloud-storage>=2.10.0
firebase-admin>=6.2.0
twilio>=8.10.0
geopy>=2.4.0
geojson>=3.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
